+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/TasksDataController/UpdateOneHundredAndTwentyFiveYearLeaseTaskTests.cs
@@ -1,0 +1,125 @@
+﻿using AutoFixture;
+using Dfe.Complete.Api.Tests.Integration.Customizations;
+using Dfe.Complete.Client.Contracts;
+using Dfe.Complete.Domain.Entities;
+using Dfe.Complete.Infrastructure.Database;
+using Dfe.Complete.Tests.Common.Constants;
+using Dfe.Complete.Utils.Exceptions;
+using GovUK.Dfe.CoreLibs.Testing.AutoFixture.Attributes;
+using GovUK.Dfe.CoreLibs.Testing.Mocks.WebApplicationFactory;
+using Microsoft.EntityFrameworkCore;
+using System.Security.Claims; 
+
+namespace Dfe.Complete.Api.Tests.Integration.Controllers.TasksDataController
+{
+    public class UpdateOneHundredAndTwentyFiveYearLeaseTaskTests
+    { 
+        [Theory]
+        [CustomAutoData(
+            typeof(CustomWebApplicationDbContextFactoryCustomization),
+            typeof(ConversionTaskDataCustomization))]
+        public async Task UpdateUpdateOneHundredAndTwentyFiveYearLeaseTaskAsync_ShouldUpdate_TransferTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+
+            var command = new UpdateOneHundredAndTwentyFiveYearLeaseTaskCommand
+            {
+                TaskDataId = new TaskDataId { Value = taskData.Id.Value }, 
+                Email = true,
+                Receive = true,
+                Save = true,
+            };
+
+            // Act
+            await tasksDataClient.UpdateUpdateOneHundredAndTwentyFiveYearLeaseTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.True(existingTaskData.OneHundredAndTwentyFiveYearLeaseEmail);
+            Assert.True(existingTaskData.OneHundredAndTwentyFiveYearLeaseReceive);
+            Assert.True(existingTaskData.OneHundredAndTwentyFiveYearLeaseSaveLease);
+            Assert.Null(existingTaskData.OneHundredAndTwentyFiveYearLeaseNotApplicable);
+        }
+        
+        
+        [Theory]
+        [CustomAutoData(
+            typeof(CustomWebApplicationDbContextFactoryCustomization),
+            typeof(ConversionTaskDataCustomization))]
+        public async Task UpdateUpdateOneHundredAndTwentyFiveYearLeaseTaskAsync_ShouldNotUpdate_IfNotApplicableIsTrue(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+
+            var command = new UpdateOneHundredAndTwentyFiveYearLeaseTaskCommand
+            {
+                TaskDataId = new TaskDataId { Value = taskData.Id.Value },
+                NotApplicable = true,
+                Email = true,
+                Receive = true,
+                Save = true,
+            };
+
+            // Act
+            await tasksDataClient.UpdateUpdateOneHundredAndTwentyFiveYearLeaseTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.Null(existingTaskData.OneHundredAndTwentyFiveYearLeaseEmail);
+            Assert.Null(existingTaskData.OneHundredAndTwentyFiveYearLeaseReceive);
+            Assert.Null(existingTaskData.OneHundredAndTwentyFiveYearLeaseSaveLease);
+            Assert.True(existingTaskData.OneHundredAndTwentyFiveYearLeaseNotApplicable);
+        }
+
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateUpdateOneHundredAndTwentyFiveYearLeaseTaskAsync_ShouldFail_WhenTaskDataIdNotMatched(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+
+            var command = new UpdateOneHundredAndTwentyFiveYearLeaseTaskCommand
+            {
+                TaskDataId = new TaskDataId { Value = Guid.NewGuid() }
+            };
+
+            // Act + Assert
+            var exception = await Assert.ThrowsAsync<NotFoundException>(() => tasksDataClient.UpdateUpdateOneHundredAndTwentyFiveYearLeaseTaskAsync(command, default));
+
+            Assert.Contains($"Conversion task data TaskDataId {{ Value = {command.TaskDataId.Value} }} not found.", exception.Message);
+        }
+    }
+}